+++ src/main/kotlin/LotteryController.kt
@@ -1,3 +1,5 @@
+import service.LotteryService
+
 class LotteryController {
 
     fun startLotteryService() {
