+++ src/main/kotlin/service/LotteryCalculator.kt
@@ -1,3 +1,5 @@
+package service
+
 import domain.Lottery
 import domain.WinningFinder
