+++ src/main/kotlin/service/Settlement.kt
@@ -1,4 +1,4 @@
-package domain
+package service
 
 object Settlement {
