+++ src/main/kotlin/util/RandomNumberGenerator.kt
@@ -1,7 +1,5 @@
 package util
 
-object RandomNumberGenerator : NumberGenerator {
-    override fun randomNumbers(): Set<Int> {
-        return (1..45).shuffled().take(6).toSet()
-    }
+val randomNumberGenerator: NumberGenerator = NumberGenerator {
+    (1..45).shuffled().take(6).toSet()
 }
